Most of the way to splitting out NameExpressionNodeVisitor
